divided the data into modules
