+++ index.html
@@ -233,6 +233,5 @@
   </template>
 
   <script type="module" src="./js/main.js"></script>
-  <script type="module" src="./js/functions.js"></script>
 </body>
 </html>
